# src/controllers/contacts.js
import {
  createContact,
  deleteContact,
  updateContact,
  getAllContacts,
  getContactById,
} from '../services/contacts.js';
import createHttpError from 'http-errors';
import { parsePaginationParams } from '../utils/parsePaginationParams.js';
import { parseSortParams } from '../utils/parseSortParams.js';
import { saveFileToCloudinary } from '../utils/saveFileToCloudinary.js';

export const createContactController = async (req, res) => {
  let photoUrl = null;

  if (req.file) {
    photoUrl = await saveFileToCloudinary(req.file);
  }

  const contact = await createContact({
    ...req.body,
    userId: req.user._id,
    photo: photoUrl,
  });

  res.status(201).json({
    status: 201,
    message: 'Successfully created a contact!',
    data: contact,
  });
};

export const getContactsController = async (req, res) => {
  const { page, perPage } = parsePaginationParams(req.query);
  const { sortBy, sortOrder } = parseSortParams(req.query);

  const contacts = await getAllContacts({
    page,
    perPage,
    sortBy,
    sortOrder,
    userId: req.user._id,
  });

  res.json({
    status: 200,
    message: 'Successfully found contacts!',
    data: contacts,
  });
};

export const getContactByIdController = async (req, res) => {
  const { contactId } = req.params;
  const contact = await getContactById(contactId, req.user._id);

  if (!contact) {
    throw createHttpError.NotFound(`Contact not found: ${contactId}`);
  }

  res.json({
    status: 200,
    message: `Successfully found contact with id ${contactId}!`,
    data: contact,
  });
};

export const deleteContactController = async (req, res) => {
  const { contactId } = req.params;
  const result = await deleteContact(contactId, req.user._id);

  if (!result) {
    throw createHttpError.NotFound('Contact not found');
  }

  res.json({
    status: 200,
    message: 'Contact deleted successfully',
  });
};

export const patchContactController = async (req, res) => {
  const { contactId } = req.params;
  let photoUrl = req.body.photo || null;

  if (req.file) {
    photoUrl = await saveFileToCloudinary(req.file);
  }

  const result = await updateContact(contactId, req.user._id, {
    ...req.body,
    photo: photoUrl,
  });

  if (!result) {
    throw createHttpError.NotFound('Contact not found');
  }

  res.json({
    status: 200,
    message: 'Successfully patched a contact!',
    data: result,
  });

  // res.status(200).json({
  //   status: 'success',
  //   message: 'Contact updated successfully',
  //   data: result,
  // });
};